Added "cv_c_schwinger" gate, and fixed yQB sign error.
Added a new gate, cv_c_schwinger. This is a very general gate that contains as a limiting case a conditional beamsplitter of arbitrary phase and qubit polarization.

Also fixed a bug in operators.py, where yQB (intended to be the Pauli-Y operator), was off by an overall minus sign. This would have impacted sign conventions for all hybrid qubit-qumode gates.